+++ gradle-plugin/src/main/kotlin/io/github/oscardlfr/kmptestrunner/tasks/CoverageTask.kt
@@ -3,12 +3,15 @@ package io.github.oscardlfr.kmptestrunner.tasks
 
 import io.github.oscardlfr.kmptestrunner.KmpTestRunnerExtension
 import org.gradle.api.DefaultTask
+import org.gradle.api.tasks.Internal
 import org.gradle.api.tasks.TaskAction
 import java.nio.file.Files
 import java.nio.file.StandardCopyOption
 
 open class CoverageTask : DefaultTask() {
+    @get:Internal
     lateinit var extension: KmpTestRunnerExtension
+    @get:Internal
     internal var koverDetected: Boolean = false
 
     @TaskAction